[Style][Ops] Share the routed-expert eager path, and name the test for its module
`MoeGateUpFwdOp` and `MoeGroupedGemmNopadFwdOp` each carried their own
`_eager_forward`, 21 lines identical character for character: both are called as
`(a, b, true_sizes, true_offsets)`, so both check the same dtypes, the same
device agreement, and resolve the kernel the same way. It is
`GroupedOperandEagerForward` in `_common.py` now, which both inherit.

Unlike `_validate_dtypes`, `_eager_forward` is no codegen slot -- nothing
installs it from the manifest -- so moving it out of the class body changes
nothing about what the class resolves. Both still take `_validate_dtypes` from
their own `__dict__`, and the mixin sits ahead of `Op` in each MRO.

`tests/kernels/test_linear_attn_autotune.py` was named for a module that the
linear-attention grouping renamed; it is `test_linear_attention_autotune.py`,
matching the `linear_attention.autotune` it imports.

Verified: the resolved `_eager_forward` has the same AST as the two copies it
replaces, docstrings aside.
