+++ src/tileops/ops/moe/routed_expert/moe_grouped_gemm_nopad.py
@@ -10,6 +10,7 @@ from tileops.kernels.moe.moe_grouped_gemm_nopad import MoeGroupedGemmNopadKernel
 
 from ...compile_boundary import get_instance
 from ...op_base import Op
+from ._common import GroupedOperandEagerForward
 
 __all__ = ["MoeGroupedGemmNopadFwdOp"]
 
@@ -17,7 +18,7 @@ __all__ = ["MoeGroupedGemmNopadFwdOp"]
 _GEMM_KEYS = ("moe_grouped_gemm_kernel", "moe_grouped_gemm_persistent_kernel")
 
 
-class MoeGroupedGemmNopadFwdOp(Op):
+class MoeGroupedGemmNopadFwdOp(GroupedOperandEagerForward, Op):
     """NT grouped GEMM for MoE without block_m-aligned padding.
 
     Uses a GPU tile scheduler to map each CTA to its (expert, row_offset) in O(1),
@@ -119,28 +120,6 @@ class MoeGroupedGemmNopadFwdOp(Op):
             C: [numel, N] GEMM output.
         """
         return _moe_grouped_gemm_nopad_fwd(a, b, true_sizes, true_offsets, self._instance_key)
-
-    def _eager_forward(
-        self,
-        a: torch.Tensor,
-        b: torch.Tensor,
-        true_sizes: torch.Tensor,
-        true_offsets: torch.Tensor,
-    ) -> torch.Tensor:
-        """Validate, normalize, resolve the kernel and launch, inside the operator.
-
-        Never traced: kernel construction enters a TileLang builder, which dynamo
-        cannot follow.
-        """
-        self._validate_dtypes(a, b, true_sizes, true_offsets)
-        for name, t in (("b", b), ("true_sizes", true_sizes), ("true_offsets", true_offsets)):
-            if t.device != a.device:
-                raise ValueError(f"{name} must be on {a.device}, got {t.device}")
-        self.dtype = a.dtype
-        # The op hands over what the manifest declares; how a kernel wants it laid
-        # out is its own business.
-        inputs = tuple(t.contiguous() for t in (a, b, true_sizes, true_offsets))
-        return self._get_kernel(inputs, a.dtype)(*inputs)
 
 
 @torch.library.custom_op("top::moe_grouped_gemm_nopad_fwd", mutates_args=())
